requirements.txt
kaggle
dagshub
dvc
dvc-http
pandas
psycopg2-binary
requests
numpy
pyspark